pyproject.toml
[project]
name = "inventory-managment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.99.9",
    "openai-agents>=0.2.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
